split out variants into own folder
