+++ static/src/javascripts/projects/common/modules/identity/sign-in-gate/component.js
@@ -1,0 +1,10 @@
+// @flow
+
+// component name, should always be sign-in-gate
+export const componentName = 'sign-in-gate';
+
+// set the ophan component tracking vars
+export const component: OphanComponent = {
+    componentType: 'SIGN_IN_GATE',
+    id: 'tertius_test',
+};
