+++ static/src/javascripts/projects/common/modules/identity/sign-in-gate/variants/index.js
@@ -1,0 +1,6 @@
+// @flow
+import type { SignInGateVariant } from '../types';
+import { signInGateVariant as control } from './control';
+import { signInGateVariant as variant } from './variant';
+
+export const variants: Array<SignInGateVariant> = [control, variant];
